pyproject.toml
[project]
name = "podcast-summary"
version = "0.1.0"
description = "Automated podcast monitoring, transcription, and summarization system"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "feedparser>=6.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "assemblyai>=0.17.0",
    "openai>=1.0.0",
    "google-genai>=0.1.0",
    "resend>=0.8.0",
    "click>=8.0.0",
    "markdown2>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
